[no jira] Make theming common amongst native stories
